Restore the test status dropdown for datatables 2
We have to query a new element id (dt-length-0 instead of results_length)
since the old one is no longer available and select/input elements
always have ID-s so they are more reliable across updates. We also
follow the approach in datatables to provide ID for the added status
select element and reuse the dt-search class for the new div element
with the only exception that we need extra right margin to distance
a colon appended "status" from the select element (only the input
element will inherit the right marging from dt-search).

Signed-off-by: Plamen Dimitrov <plamen.dimitrov@intra2net.com>