Restore resizer div height and thus the overall column resizing
We have to remove the static div elements in the HTML template because
datatables will include all of them in a span for the header cell
content, overwriting the original table structure, and must instead
perform our own overwriting afterwards. The div elements are added to
each table header cell if they are not already present to add them
just once for all table redraws.

If the div is instead wrapped in a span by datatables this will lead
to its dynamically calculated height to be added to ever increasing
table header cell height and degrade the table headers. The current
change achieves the same structure we had before but with multiple
spans (for content and for sorting arrows) now with datatables 2.

Signed-off-by: Plamen Dimitrov <plamen.dimitrov@intra2net.com>